Rename 'components/DataTable.vue' to 'components/index/CardsReference/ConfirmedCasesAttributes/DataTable.vue'
